feat: get user orders

+++ internal/models/order.go
@@ -17,7 +17,7 @@ type Order struct {
 	Number     string      `json:"number" db:"number"`
 	UserID     string      `json:"-" db:"user_id"`
 	Status     OrderStatus `json:"status" db:"status"`
-	Accrual    *float64    `json:"accrual,omitempty" db:"accrual"`
+	Accrual    *int        `json:"accrual,omitempty" db:"accrual"`
 	UploadedAt time.Time   `json:"uploaded_at" db:"uploaded_at"`
 }
